feat!: improve the command line interface
The old behavior is always running the command as a daemon.
The new behavior is to run the command once and exit if the first argument
is not `--daemon`.
And if the first argument is `--daemon`, it will run as a daemon.

Besides, if the command is not specified, it will try to run the
`$HOME/.config/dark-mode-notify/notify` script.

BREAKING CHANGE: The behavior of the command line interface has changed.